feat: 통계 페이지 기본 ui
