+++ src/Routing.tsx
@@ -1,5 +1,7 @@
 // src/Routing.tsx
 import { Routes, Route } from "react-router-dom";
+import { useAuthStore } from "./stores/useAuthStore";
+
 import Layout from "@/components/Layout";
 import Login from "./pages/Login";
 import About from "./pages/About";
@@ -11,8 +13,7 @@ import Emulator from "./pages/emulator/Emulator";
 import PrivateRoute from "./components/PrivateRoute";
 import Register from "./pages/Register";
 import DashboardMain from "./pages/dashboard/DashBoardMain";
-// import Home from "./pages/Home";
-import { useAuthStore } from "./stores/useAuthStore";
+
 import HistoryMain from "./pages/history/HistoryMain";
 import InfoMain from "./pages/info/InfoMain";
 import AdminMain from "./pages/admin/AdminMain";
@@ -20,6 +21,7 @@ import AdminRoute from "./components/AdminRoute";
 import QuestMain from "./pages/quest/QuestMain";
 import HistoryRealTime from "./pages/history/realtime/HistoryRealTime";
 import HistoryDetailPage from "./pages/history/detail/HistoryDetailPage";
+import StatisticMain from "./pages/statistic/StatisticMain";
 
 // 로그인 없이 접근 가능한 라우트들
 const publicRoutes = [
@@ -33,13 +35,14 @@ const publicRoutes = [
 // 로그인 후에만 접근 가능한 라우트들
 const protectedRoutes = [
   { path: "/dashboard", element: <DashboardMain /> },
-  { path: "/cars", element: <CarMain /> },
-  { path: "/cars/register", element: <CarRegister /> },
-  { path: "/rents", element: <RentMain /> },
-  { path: "/rents/register", element: <RentRegister /> },
+  { path: "/car", element: <CarMain /> },
+  { path: "/car/register", element: <CarRegister /> },
+  { path: "/car/rent", element: <RentMain /> },
+  { path: "/car/rent/register", element: <RentRegister /> },
   { path: "/history", element: <HistoryMain /> },
   { path: "/history/:id", element: <HistoryDetailPage /> },
   { path: "/history/realtime", element: <HistoryRealTime /> },
+  { path: "/statistic", element: <StatisticMain /> },
   { path: "/quest", element: <QuestMain /> },
 ];
