+++ src/pages/statistic/StatisticCarSection.tsx
@@ -1,0 +1,9 @@
+function StatisticCarSection() {
+  return (
+    <div className="w-full h-full flex flex-col p-10">
+      차량 통계 표시
+    </div>
+  );
+}
+
+export default StatisticCarSection;
