+++ src/pages/statistic/daily/StatisticDailySection.tsx
@@ -1,0 +1,5 @@
+function StatisticDailySection() {
+  return <div>StatisticDailySection</div>;
+}
+
+export default StatisticDailySection;
