+++ src/pages/statistic/monthly/StatisticMonthSection.tsx
@@ -1,0 +1,5 @@
+function StatisticMonthSection() {
+  return <div>StatisticMonthSection</div>;
+}
+
+export default StatisticMonthSection;
